feat: refactor authentication logic (#103)
