+++ apps/web/app/(app)/profile/[token]/route.ts
@@ -8,17 +8,19 @@ import {
 	incrementShareableLinkUsage,
 } from "~/features/shareable-links/data/shareable-links";
 
-export default async function GET(
+export async function GET(
 	_request: Request,
 	{ params }: { params: Promise<{ token: string }> },
 ) {
 	const { token } = await params;
+
 	const link = await getShareableLinkByToken(token);
+	console.log(link);
 
 	if (!link || link.revoked) return notFound();
 	if (link.expiresAt && new Date(link.expiresAt) < new Date())
 		return notFound();
-	if (link.usageLimit !== null && link.usageCount >= link.usageLimit)
+	if (link.usageLimit !== 0 && link.usageCount >= link.usageLimit)
 		return notFound();
 
 	after(async () => await incrementShareableLinkUsage(token));
