+++ packages/auth/src/auth.config.ts
@@ -2,8 +2,6 @@ import type { NextAuthConfig } from "next-auth";
 import Credentials from "next-auth/providers/credentials";
 import Spotify from "next-auth/providers/spotify";
 
-import { getUserByShareableLinkToken } from "./queries";
-
 export default {
 	providers: [
 		Spotify({
@@ -11,6 +9,7 @@ export default {
 				"https://accounts.spotify.com/authorize?scope=user-read-recently-played%20user-top-read%20user-read-email",
 		}),
 		Credentials({
+			id: "demo",
 			name: "Demo",
 			credentials: {
 				username: { label: "Username", type: "text" },
@@ -29,26 +28,6 @@ export default {
 					};
 				}
 				return null;
-			},
-		}),
-		Credentials({
-			name: "Shareable Link",
-			credentials: {
-				token: { label: "Token", type: "string" },
-			},
-			async authorize(credentials) {
-				const data = await getUserByShareableLinkToken(
-					credentials.token as string,
-				);
-
-				if (!data) return null;
-
-				return {
-					name: "Shareable Link",
-					email: "demo@demo.com",
-					id: data.User.id,
-					hasPackage: data.User.hasPackage ?? false,
-				};
 			},
 		}),
 	],
