+++ packages/auth/src/auth.ts
@@ -1,9 +1,11 @@
 import { DrizzleAdapter } from "@auth/drizzle-adapter";
 import NextAuth, { type DefaultSession, type NextAuthResult } from "next-auth";
+import Credentials from "next-auth/providers/credentials";
 
 import { accounts, db, users } from "@repo/database";
 
 import authConfig from "./auth.config";
+import { getUserByShareableLinkToken } from "./queries";
 
 declare module "next-auth" {
 	interface Session {
@@ -18,6 +20,7 @@ declare module "next-auth" {
 }
 
 const result = NextAuth({
+	...authConfig,
 	adapter: {
 		...DrizzleAdapter(db, {
 			usersTable: users,
@@ -25,7 +28,30 @@ const result = NextAuth({
 		}),
 	},
 	session: { strategy: "jwt" },
-	...authConfig,
+	providers: [
+		...authConfig.providers,
+		Credentials({
+			id: "shareable-link",
+			name: "Shareable Link",
+			credentials: {
+				token: { label: "Token", type: "string" },
+			},
+			async authorize(credentials) {
+				const data = await getUserByShareableLinkToken(
+					credentials.token as string,
+				);
+
+				if (!data) return null;
+
+				return {
+					name: "Shareable Link",
+					email: "demo@demo.com",
+					id: data.User.id,
+					hasPackage: data.User.hasPackage ?? false,
+				};
+			},
+		}),
+	],
 });
 
 export const handlers: NextAuthResult["handlers"] = result.handlers;
